api/requirements.txt
flask
requests
bs4
sklearn
pandas
numpy
google.cloud
google-cloud-storage
google-cloud-language